Uploaded Action File
This will be used on all of the robots so both programs have a similar source, and so our OP modes will be clear. Just as a note for newbies, there are somethings in this file that look really confusing to a new programmer, just know that as of now it really shouldn't matter if you know what the methods do or not.